Differ between option sets and structures with optional fields
According to the spec part 3 (see e.g. 8.41) the `IsOptionSet`
attribute has nothing in common with structures with optional
fields. This was a misunderstanding due to the naming. An
option set is used for bit mask representation.

With this commit structures with optional fields are correctly
identified and option sets get their own record. The
serialization of option sets is NOT done within this commit,
only the schemas are aligned so far.